refactor(docker/containers): parse view model
