+++ app/react/docker/containers/CreateView/CommandsTab/index.ts
@@ -1,5 +1,14 @@
+import { validation } from './validation';
+import { toRequest } from './toRequest';
+import { toViewModel, getDefaultViewModel } from './toViewModel';
+
 export { CommandsTab } from './CommandsTab';
 export { validation as commandsTabValidation } from './validation';
-export { parseRequest as parseCommandsTabRequest } from './parseRequest';
-export { parseViewModel as parseCommandsTabViewModel } from './parseViewModel';
 export { type Values as CommandsTabValues } from './types';
+
+export const commandsTabUtils = {
+  toRequest,
+  toViewModel,
+  validation,
+  getDefaultViewModel,
+};
