--- app/react/docker/containers/CreateView/CommandsTab/parseViewModel.tsx
@@ -1,66 +1,0 @@
-import { HostConfig } from 'docker-types/generated/1.41';
-
-import { commandArrayToString } from '@/docker/helpers/containers';
-
-import { ContainerJSON } from '../../queries/container';
-
-import { ConsoleConfig, ConsoleSetting } from './ConsoleSettings';
-import { LogConfig } from './LoggerConfig';
-import { Values } from './types';
-
-export function parseViewModel(config?: ContainerJSON): Values {
-  if (!config || !config.Config || !config.HostConfig) {
-    return {
-      cmd: null,
-      entrypoint: null,
-      user: '',
-      workingDir: '',
-      console: 'none',
-      logConfig: getLogConfig(),
-    };
-  }
-
-  return {
-    cmd: config.Config.Cmd ? commandArrayToString(config.Config.Cmd) : null,
-    entrypoint: config.Config.Entrypoint
-      ? commandArrayToString(config.Config.Entrypoint)
-      : null,
-    user: config.Config.User || '',
-    workingDir: config.Config.WorkingDir || '',
-    console: config ? getConsoleSetting(config.Config) : 'none',
-    logConfig: getLogConfig(config.HostConfig.LogConfig),
-  };
-
-  function getLogConfig(value?: HostConfig['LogConfig']): LogConfig {
-    if (!value || !value.Type) {
-      return {
-        type: 'none',
-        options: [],
-      };
-    }
-
-    return {
-      type: value.Type,
-      options: Object.entries(value.Config || {}).map(([option, value]) => ({
-        option,
-        value,
-      })),
-    };
-  }
-
-  function getConsoleSetting(value: ConsoleConfig): ConsoleSetting {
-    if (value.OpenStdin && value.Tty) {
-      return 'both';
-    }
-
-    if (!value.OpenStdin && value.Tty) {
-      return 'tty';
-    }
-
-    if (value.OpenStdin && !value.Tty) {
-      return 'interactive';
-    }
-
-    return 'none';
-  }
-}
